refactor: 각 Json 요청을 매핑해주던 Mapper들 JsonRequestMapper로 통합, EqualOwner객체 내에서 동일 Owner인지 점검하도록 수정
