+++ src/main/java/toolc/yourlist/playlist/domain/EqualOwner.java
@@ -1,4 +1,9 @@
 package toolc.yourlist.playlist.domain;
 
 public record EqualOwner(Member member, Playlist playlist) {
+  private final static EqualOwnerCondition condition = new EqualOwnerCondition();
+
+  public EqualOwner {
+    condition.check(member, playlist);
+  }
 }
