+++ src/main/java/toolc/yourlist/playlist/domain/EqualOwnerCondition.java
@@ -1,8 +1,5 @@
 package toolc.yourlist.playlist.domain;
 
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
 final class EqualOwnerCondition {
   void check(Member member, Playlist playlist) {
     if (!member.id().equals(playlist.memberId())) {
