+++ src/main/java/toolc/yourlist/playlist/domain/SaveRequest.java
@@ -1,4 +1,4 @@
 package toolc.yourlist.playlist.domain;
 
-public record SaveRequest(Member member, String title) {
+public record SaveRequest(Long memberId, String title) {
 }
